Added ExponentialCallback for trian.Loop.
